ingestors/fred/pyproject.toml
[project]
name = "fred"
version = "0.1.0"
description = "Pulls data from FRED public datasets"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fredapi>=0.5.2",
    "google-cloud-storage>=3.8.0",
    "pandas>=2.3.3",
]
